refactor spell visual code
